requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
requests==2.18.4
requests-oauthlib==0.8.0
selenium==3.6.0
simplediff==1.0
six==1.11.0
SQLAlchemy==1.1.13
tweepy==3.5.0
urllib3==1.22
